backend/requirements.txt
Django==3.0.5
djangorestframework==3.12.2
elasticsearch-dsl==7.4.0
pyre2==0.3.5
django-celery-beat==2.2.0
celery==5.0.5
django-celery-results==2.0.1
psycopg2==2.8.6
django-cors-headers==3.7.0